Move connectors detection to gpu.cpp
